requirements.txt
deepspeed==0.14.0
jsonlines==4.0.0
numpy==1.25.2
pandas==2.2.3
Pillow==11.0.0
PyYAML==6.0.2
PyYAML==6.0.2
Requests==2.32.3
scikit_learn==1.2.2
scipy==1.14.1
torch==2.4.0
torch==2.3.1
tqdm==4.66.5
transformers==4.44.2
